requirements.txt
amqp==5.0.3
anyjson==0.3.3
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
django-cors-headers==3.6.0
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.20.9
django-nine==0.2.3
djangorestframework==3.12.2
drf-yasg==1.20.0
elasticsearch==7.10.0
elasticsearch-dsl==7.3.0
et-xmlfile==1.0.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
openpyxl==3.0.5
packaging==20.4
prompt-toolkit==3.0.11
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
